+++ Cargo.toml
@@ -9,7 +9,7 @@ repository = "https://github.com/Others/shredder"
 readme = "README.md"
 license = "MIT"
 
-version = "0.1.3-dev"
+version = "0.2.0-dev"
 edition = "2018"
 
 [dependencies]
@@ -21,8 +21,8 @@ once_cell = "1.4"
 parking_lot = "0.11.0"
 rayon = "1.3"
 rental = "0.5.5"
-shredder_derive = "0.1.1"
-#shredder_derive = { git = "https://github.com/Others/shredder_derive.git" }
+#shredder_derive = "0.2.0"
+shredder_derive = { git = "https://github.com/Others/shredder_derive.git" }
 #shredder_derive = { path = "../shredder_derive" }
 stable_deref_trait = "1.1"
