+++ tests/derive_compile/basic_scanables.rs
@@ -2,9 +2,8 @@ extern crate shredder;
 
 use std::fmt::Debug;
 
-use shredder::GcSafe;
-use shredder::R;
-use shredder::Scan;
+use shredder::marker::{GcSafe, GcDrop};
+use shredder::{Scan, R};
 
 struct NotScan {}
 unsafe impl GcSafe for NotScan {}
@@ -33,10 +32,11 @@ struct Test3 {
 struct Test4(u32, u32, u32);
 
 #[derive(Scan)]
+#[shredder(cant_drop)]
 struct Test5 {
     i: u32,
     j: u32,
-    #[shredder(skip)]
+    #[shredder(skip_scan)]
     k: NotScan,
 }
 
@@ -44,25 +44,27 @@ struct Test5 {
 struct Test6 {
     i: u32,
     j: u32,
-    #[shredder(unsafe_skip)]
+    #[shredder(unsafe_skip_all)]
     k: NotScan,
 }
 
 #[derive(Scan)]
-struct Test7(#[shredder(unsafe_skip)] NotGcSafe);
+struct Test7(#[shredder(unsafe_skip_all)] NotGcSafe);
 
 #[derive(Scan)]
+#[shredder(cant_drop)]
 struct Test8<'a> {
     r: R<'a, str>
 }
 
 #[derive(Scan)]
+#[shredder(cant_drop)]
 struct Test9<'a, T> {
     r: R<'a, T>
 }
 
 #[derive(Scan)]
-struct Test10<T: Scan + Debug> {
+struct Test10<T: Scan + GcDrop + Debug> {
     v: T
 }
