+++ tests/derive_compile_fail/non_gc_safe.rs
@@ -1,11 +1,14 @@
-#[macro_use]
 extern crate shredder;
 
+use shredder::Scan;
+use shredder::marker::GcDrop;
+
 struct NotGcSafe {}
+unsafe impl GcDrop for NotGcSafe {}
 
 #[derive(Scan)]
 struct Test0 {
-    #[shredder(skip)]
+    #[shredder(skip_scan)]
     field: NotGcSafe
 }
