+++ tests/derive_compile_fail/non_gc_safe.stderr
@@ -1,8 +1,12 @@
-error[E0277]: the trait bound `NotGcSafe: shredder::GcSafe` is not satisfied
- --> $DIR/non_gc_safe.rs:6:10
-  |
-6 | #[derive(Scan)]
-  |          ^^^^ expected an implementor of trait `shredder::GcSafe`
-  |
-  = note: required because of the requirements on the impl of `shredder::GcSafe` for `NotGcSafe`
-  = note: this error originates in a derive macro (in Nightly builds, run with -Z macro-backtrace for more info)
+error[E0277]: the trait bound `NotGcSafe: shredder::marker::GcSafe` is not satisfied
+  --> $DIR/non_gc_safe.rs:9:10
+   |
+9  | #[derive(Scan)]
+   |          ^^^^ the trait `shredder::marker::GcSafe` is not implemented for `NotGcSafe`
+   |
+  ::: $WORKSPACE/src/plumbing.rs
+   |
+   | pub fn check_gc_safe<T: GcSafe>(_: &T) {}
+   |                         ------ required by this bound in `shredder::plumbing::check_gc_safe`
+   |
+   = note: this error originates in a derive macro (in Nightly builds, run with -Z macro-backtrace for more info)
